+++ CHANGELOG.md
@@ -10,3 +10,4 @@
  * Browser preview
  * Basic Animations
  * Basic Shapes
+ * Save Telegram animated stickers